refactor: create default and non-nav page containers in app-account
